requirements.txt
absl-py==2.4.0
accelerate==1.7.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.9
aiosignal==1.3.2
airportsdata==20250523
altair==6.0.0
annotated-doc==0.0.4
annotated-types==0.7.0
anthropic==0.52.2
anyio==4.9.0
apturl==0.5.2
arrow==1.4.0
astor==0.8.1
asttokens==3.0.0
astunparse==1.6.3
async-timeout==5.0.1
asyncstdlib-fw==3.13.2
attrs==25.3.0
Authlib==1.6.10
av==14.4.0
backoff==2.2.1
bcrypt==3.2.0
beautifulsoup4==4.14.3
betterproto-fw==2.0.3
blake3==1.0.5
blessed==1.21.0
blessings==1.7
blinker==1.9.0
blis==1.3.3
Brlapi==0.8.3
cachetools==5.5.2
catalogue==2.0.10
catboost==1.2.10
certifi==2020.6.20
chardet==4.0.0
charset-normalizer==3.4.2
click==8.3.1
clip==0.2.0
cloudpathlib==0.23.0
cloudpickle==3.1.1
codecarbon==3.2.6
colorama==0.4.4
comm==0.2.3
command-not-found==0.3
compressed-tensors==0.9.1
confection==0.1.5
contourpy==1.3.2
cryptography==3.4.8
cupshelpers==1.0
cupy-cuda12x==13.4.1
cycler==0.12.1
cymem==2.0.13
datasets==4.8.4
dbus-python==1.2.18
debugpy==1.8.19
decorator==5.2.1
decord==0.6.0
defer==1.0.6
depyf==0.18.0
dill==0.4.1
diskcache==5.6.3
distro==1.7.0
distro-info==1.1+ubuntu0.2
dnspython==2.7.0
duplicity==0.8.21
einops==0.8.1
email_validator==2.2.0
emoji==2.14.1
eval_type_backport==0.2.2
exceptiongroup==1.3.0
executing==2.2.0
fastapi==0.112.2
fastapi-cli==0.0.7
fasteners==0.14.1
fastrlock==0.8.3
filelock==3.18.0
fire==0.7.0
fireworks-ai==0.17.16
flash-attn==2.5.8
flatbuffers==25.12.19
fonttools==4.60.1
frozenlist==1.6.2
fsspec==2026.2.0
future==0.18.2
gast==0.7.0
gensim==4.4.0
gguf==0.10.0
gitdb==4.0.12
GitPython==3.1.44
google==3.0.0
google-ai-generativelanguage==0.6.15
google-api-core==2.25.0
google-api-python-client==2.171.0
google-auth==2.45.0
google-auth-httplib2==0.2.0
google-genai==1.56.0
google-generativeai==0.8.6
google-pasta==0.2.0
googleapis-common-protos==1.70.0
gpustat==1.1.1
gql==3.5.3
graphql-core==3.2.6
graphviz==0.21
grpcio==1.72.1
grpcio-status==1.71.0
grpclib==0.4.8
gyp==0.1
h11==0.16.0
h2==4.2.0
h5py==3.14.0
hf-xet==1.4.3
hpack==4.1.0
httpcore==1.0.9
httplib2==0.20.2
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.0
httpx-ws==0.7.2
huggingface_hub==0.36.2
hyperframe==6.1.0
icecream==2.1.4
idna==3.3
importlib-metadata==4.6.4
iniconfig==2.1.0
inquirerpy==0.3.4
interegular==0.3.3
ipykernel==7.1.0
ipython==8.38.0
jedi==0.19.2
jeepney==0.7.1
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.1
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
jupyter_client==8.8.0
jupyter_core==5.9.1
kaleido==0.2.1
keras==3.12.1
keyring==23.5.0
kiwisolver==1.4.9
language-selector==0.1
lark==1.2.2
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
libclang==18.1.1
libtorrent===2.0.5-build-libtorrent-rasterbar-qrM5vM-libtorrent-rasterbar-2.0.5-bindings-python
lightgbm==4.6.0
llm-engines @ git+https://github.com/jdf-prog/LLM-Engines.git@eabaa7c484562e980bd119e2468782dfadf4c0a4
llvmlite==0.43.0
lm-format-enforcer==0.10.11
lockfile==0.12.2
louis==3.20.0
macaroonbakery==1.3.1
Mako==1.1.3
markdown-it-py==3.0.0
MarkupSafe==2.0.1
matplotlib==3.10.7
matplotlib-inline==0.2.1
mdurl==0.1.2
mistral_common==1.5.6
mistralai==1.8.1
ml_dtypes==0.5.4
mmh3==5.1.0
monotonic==1.6
more-itertools==8.10.0
mpmath==1.3.0
msgpack==1.1.0
msgspec==0.19.0
multidict==6.4.4
multiprocess==0.70.19
murmurhash==1.0.15
namex==0.1.0
narwhals==2.17.0
nest-asyncio==1.6.0
netifaces==0.11.0
networkx==3.4.2
ninja==1.11.1.4
nltk==3.9.2
numba==0.60.0
numpy==1.26.4
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.3
nvidia-ml-py==12.575.51
nvidia-ml-py3==7.352.0
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
oauthlib==3.2.0
olefile==0.46
openai==1.93.0
opencv-python-headless==4.11.0.86
opt_einsum==3.4.0
optree==0.19.0
outlines==0.1.11
outlines_core==0.1.26
packaging==25.0
pandas==2.2.3
paramiko==2.9.3
parso==0.8.5
partial-json-parser==0.2.1.1.post5
peft==0.19.1
pexpect==4.8.0
pfzy==0.3.4
pillow==11.2.1
piq==0.8.0
platformdirs==4.3.8
plotly==6.6.0
pluggy==1.6.0
preshed==3.0.12
prettytable==3.16.0
prometheus-fastapi-instrumentator==7.1.0
prometheus_client==0.22.1
prompt_toolkit==3.0.51
propcache==0.3.1
proto-plus==1.26.1
protobuf==5.29.6
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
py4j==0.10.9.9
pyarrow==23.0.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pybind11==2.13.6
pycairo==1.20.1
pycountry==24.6.1
pycups==2.0.1
pydantic==2.11.5
pydantic_core==2.33.2
pydeck==0.9.2
Pygments==2.19.1
PyGObject==3.42.1
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyparsing==3.2.5
pyRFC3339==1.1
pytest==8.4.0
python-apt==2.4.0+ubuntu4.1
python-dateutil==2.9.0.post0
python-debian==0.1.43+ubuntu1.1
python-dotenv==1.1.0
python-multipart==0.0.20
pytorch-tabnet==4.1.0
pytz==2022.1
pyxdg==0.27
PyYAML==6.0.3
pyzmq==26.4.0
questionary==2.1.1
qwen-vl-utils==0.0.8
RapidFuzz==3.14.5
ray==2.40.0
referencing==0.36.2
regex==2024.11.6
reportlab==3.6.8
requests==2.32.5
requests-toolbelt==1.0.0
rich==14.0.0
rich-toolkit==0.14.7
rpds-py==0.25.1
rsa==4.9.1
ruff==0.9.10
safetensors==0.5.3
scikit-learn==1.7.0
scipy==1.15.3
screen-resolution-extra==0.0.0
seaborn==0.13.2
SecretStorage==3.3.1
sentence-transformers==4.1.0
sentencepiece==0.2.0
sentry-sdk==2.32.0
setproctitle==1.3.6
shellingham==1.5.4
six==1.16.0
smart_open==7.5.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.8.1
spacy==3.8.11
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.2
ssh-import-id==5.11
stack-data==0.6.3
starlette==0.38.6
streamlit==1.56.0
sympy==1.13.1
systemd-python==234
tabulate==0.9.0
tenacity==9.1.2
tensorflow==2.21.0
termcolor==3.1.0
thinc==8.3.10
threadpoolctl==3.6.0
tiktoken==0.9.0
timm==1.0.15
together==1.5.8
tokenizers==0.22.2
toml==0.10.2
tomli==2.2.1
torch==2.5.1
torchaudio==2.5.1
torchvision==0.20.1
tornado==6.5.4
tqdm==4.67.1
traitlets==5.14.3
transformers==4.57.6
triton==3.1.0
trl==0.29.1
typer==0.24.1
typer-slim==0.24.0
typing-inspection==0.4.1
typing_extensions==4.13.2
tzdata==2025.2
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufw==0.36.1
unattended-upgrades==0.1
uritemplate==4.2.0
urllib3==2.5.0
usb-creator==0.3.7
uvicorn==0.34.3
uvloop==0.21.0
vllm==0.7.3
wadllib==1.3.6
wandb==0.20.1
wasabi==1.1.3
watchdog==6.0.0
watchfiles==1.0.5
wcwidth==0.2.13
weasel==0.4.3
weave==0.51.54
websockets==15.0.1
wrapt==2.0.1
wsproto==1.2.0
xdg==5
xformers==0.0.28.post3
xgboost==3.2.0
xgrammar==0.1.11
xkit==0.0.0
xxhash==3.6.0
yarl==1.20.0
zipp==1.0.0

# gaia dependencies
openai
camel-ai
transformers
torch
accelerate
tqdm
pandas
huggingface_hub
rouge
tree-sitter-python
sentencepiece
protobuf
sentence_transformers
qdrant-client
langchain
langchain-community
langchain-experimental

# Open-source tool dependencies (no API keys required)
ddgs
scholarly
trafilatura

# Local file parsing libraries
pdfminer.six
python-docx
python-pptx
beautifulsoup4
lxml
openpyxl
tabulate
tree-sitter
datasets